requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
optuna>=3.4.0
scikit-learn>=1.3.0
plotly>=5.17.0
openpyxl>=3.1.0
jsonschema>=4.19.0
great-expectations>=0.18.0
seaborn>=0.12.0
matplotlib>=3.7.0
scipy>=1.11.0
xgboost>=1.7.0
lightgbm>=4.0.0
